+++ packages/eas-cli/src/commands/env/__tests__/EnvPush.test.ts
@@ -1,7 +1,7 @@
 import { Config } from '@oclif/core';
 import * as fs from 'fs-extra';
 
-import { EnvironmentVariableEnvironment } from '../../../build/utils/environment';
+import { DefaultEnvironment } from '../../../build/utils/environment';
 import { ExpoGraphqlClient } from '../../../commandUtils/context/contextUtils/createGraphqlClient';
 import { testProjectId } from '../../../credentials/__tests__/fixtures-constants';
 import {
@@ -65,13 +65,13 @@ SECRET_KEY=super-secret-key`;
 
     expect(EnvironmentVariablesQuery.byAppIdAsync).toHaveBeenCalledWith(graphqlClient, {
       appId: testProjectId,
-      environment: EnvironmentVariableEnvironment.Development,
+      environment: DefaultEnvironment.Development,
       filterNames: ['EXPO_PUBLIC_API_URL', 'VARIABLE_NAME', 'SECRET_KEY'],
     });
   });
 
-  it('rejects invalid environment when using positional argument', async () => {
-    const command = new EnvPush(['invalid', '--path', testEnvPath], mockConfig);
+  it('accepts custom environment', async () => {
+    const command = new EnvPush(['custom-environment', '--path', testEnvPath], mockConfig);
 
     // @ts-expect-error
     jest.spyOn(command, 'getContextAsync').mockReturnValue({
@@ -80,9 +80,13 @@ SECRET_KEY=super-secret-key`;
       projectDir: testProjectDir,
     });
 
-    await expect(command.runAsync()).rejects.toThrow(
-      "Invalid environment. Use one of 'production', 'preview', or 'development'."
-    );
+    await command.runAsync();
+
+    expect(EnvironmentVariablesQuery.byAppIdAsync).toHaveBeenCalledWith(graphqlClient, {
+      appId: testProjectId,
+      environment: 'custom-environment',
+      filterNames: ['EXPO_PUBLIC_API_URL', 'VARIABLE_NAME', 'SECRET_KEY'],
+    });
   });
 
   describe('--force option', () => {
@@ -91,7 +95,7 @@ SECRET_KEY=super-secret-key`;
         id: 'var1',
         name: 'VARIABLE_NAME',
         value: 'old variable value',
-        environments: [EnvironmentVariableEnvironment.Development],
+        environments: [DefaultEnvironment.Development],
         createdAt: new Date().toISOString(),
         updatedAt: new Date().toISOString(),
         scope: EnvironmentVariableScope.Project,
@@ -129,7 +133,7 @@ SECRET_KEY=super-secret-key`;
           expect.objectContaining({
             name: 'VARIABLE_NAME',
             value: 'variable value',
-            environments: [EnvironmentVariableEnvironment.Development],
+            environments: [DefaultEnvironment.Development],
             overwrite: true,
           }),
         ]),
@@ -142,7 +146,7 @@ SECRET_KEY=super-secret-key`;
         id: 'var1',
         name: 'SECRET_KEY',
         value: 'old-secret-key',
-        environments: [EnvironmentVariableEnvironment.Development],
+        environments: [DefaultEnvironment.Development],
         createdAt: new Date().toISOString(),
         updatedAt: new Date().toISOString(),
         scope: EnvironmentVariableScope.Project,
@@ -182,7 +186,7 @@ SECRET_KEY=super-secret-key`;
           expect.objectContaining({
             name: 'SECRET_KEY',
             value: 'super-secret-key',
-            environments: [EnvironmentVariableEnvironment.Development],
+            environments: [DefaultEnvironment.Development],
             overwrite: true,
           }),
         ]),
@@ -195,7 +199,7 @@ SECRET_KEY=super-secret-key`;
         id: 'var1',
         name: 'VARIABLE_NAME',
         value: 'old variable value',
-        environments: [EnvironmentVariableEnvironment.Development],
+        environments: [DefaultEnvironment.Development],
         createdAt: new Date().toISOString(),
         updatedAt: new Date().toISOString(),
         scope: EnvironmentVariableScope.Project,
@@ -230,7 +234,7 @@ SECRET_KEY=super-secret-key`;
         id: 'var1',
         name: 'VARIABLE_NAME',
         value: 'old variable value',
-        environments: [EnvironmentVariableEnvironment.Development],
+        environments: [DefaultEnvironment.Development],
         createdAt: new Date().toISOString(),
         updatedAt: new Date().toISOString(),
         scope: EnvironmentVariableScope.Project,